Revert "In Progress: Setting up Networking VR"
This reverts commit 20420bb6622f9fba1ffd7cb732521d0fdbe823bd.
