+++ Assets/Prefabs.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: b20a8a894c406d244bd3681f97bc190c
-PrefabImporter:
+guid: d7265859a72af4e2182231e26950f3b1
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
