--- Assets/Project/Prefabs/Managers/GameManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0e6cdec6903c05742bacc021afd13b80
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
