--- Assets/Project/Prefabs/Managers/XRManager.prefab
@@ -1,56 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1 &735210147533532919
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 735210147533532912}
-  - component: {fileID: 735210147533532913}
-  m_Layer: 0
-  m_Name: XRManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &735210147533532912
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 735210147533532919}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &735210147533532913
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 735210147533532919}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c80fd6517612f4d4796d3bc7a7cf6f0e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  XRPlayerRig: {fileID: 0}
-  LeftHandController: {fileID: 0}
-  RightHandController: {fileID: 0}
-  LeftTeleportRay: {fileID: 0}
-  RightTeleportRay: {fileID: 0}
-  VRObjects: {fileID: 0}
-  FallbackObjects: {fileID: 0}
-  HeadCollider: {fileID: 0}
-  Teleporting: {fileID: 0}
-  XRTeleportationObject: {fileID: 0}
-  SteamTeleportationObject: {fileID: 0}
