--- Assets/Project/Prefabs/Managers/XRManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f44e99e333210e2499513bf640ffa6e4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
