+++ Assets/Project/Prefabs/Player/ServerSidePlayer.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 4483345720666607777}
-  - component: {fileID: -8280672598954856710}
+  - component: {fileID: 6108970349889774463}
   - component: {fileID: 1037102014086549081}
   m_Layer: 0
   m_Name: ServerSidePlayer
@@ -33,7 +33,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &-8280672598954856710
+--- !u!114 &6108970349889774463
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -42,16 +42,15 @@ MonoBehaviour:
   m_GameObject: {fileID: 2329821471381548265}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 54ee04d6dfb06914e9b5b93f331b770d, type: 3}
+  m_Script: {fileID: 11500000, guid: 9efb3adc5a4d54418aca2d127d6eb903, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   id: 0
   username: 
-  controller: {fileID: 0}
-  gravity_per_tick: -19.62
+  controller: {fileID: 1037102014086549081}
+  gravity_per_tick: -9.81
   moveSpeed_per_tick: 5
-  jumpSpeed_per_tick: 9
-  inputs: 0000000000
+  jumpSpeed_per_tick: 5
 --- !u!143 &1037102014086549081
 CharacterController:
   m_ObjectHideFlags: 0
