--- Assets/Project/Prefabs/Player/ServerSidePlayerVR.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d5276bc76c9469144bd7377a7f8ef799
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
