--- Assets/Project/Scenes/OperatingRoomDemo-Multiplayer.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c5f0f7bcc9f486545b3941f81e120036
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
