--- Assets/Project/Scenes/OperatingRoomDemo-Server.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b662ef4d59b449345af9bd0840df58b7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
