--- Assets/Project/Scripts/Managers/MenuManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db9b3d19252914f409a11157a7275936
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
