--- Assets/Project/Scripts/Menus.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bbe92a5c08b184c4fbac32fb306857da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
