--- Assets/Project/Scripts/Menus/Menu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c448f9289f7b8d04fbc73ae12ad18fbd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
