--- Assets/Project/Scripts/Menus/NetworkOverlayMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ead13c50dc62d145be363f83f314bfe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
