+++ Assets/Project/Scripts/NetworkingScripts/Client/GameManager.cs
@@ -15,7 +15,6 @@ namespace CEMSIM
                 // store all information about all players in game
                 public static Dictionary<int, PlayerManager> players = new Dictionary<int, PlayerManager>();
 
-                public GameObject localPlayerVR;
                 public GameObject localPlayerPrefab;
                 public GameObject playerPrefab;
 
@@ -47,14 +46,7 @@ namespace CEMSIM
                     if (_id == Client.instance.myId)
                     {
                         // create player for client
-                        if(localPlayerVR!=null)
-                        {
-                            _player = localPlayerVR;
-                        }
-                        else
-                        {
-                            _player = Instantiate(localPlayerPrefab, _position, _rotation);
-                        }                      
+                        _player = Instantiate(localPlayerPrefab, _position, _rotation);
                     }
                     else
                     {
