--- Assets/Project/Scripts/NetworkingScripts/Client/PlayerVRController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5210885a1394260438a072cf2b92e335
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
