+++ Assets/Project/Scripts/NetworkingScripts/Client/UIManager.cs
@@ -76,10 +76,7 @@ namespace CEMSIM
                     // we use Player + id to temporarily represent the player username
                     string _username = "Player" + Client.instance.myId.ToString();
 
-                    //TO DO: The Network Demo scene only works in desktop mode at the moment. 
-                    //Add feature for entering in VR or desktop mode
-                    ClientSend.SendSpawnRequest(_username, false);
-                    //ClientSend.SendSpawnRequest(_username);
+                    ClientSend.SendSpawnRequest(_username);
 
                 }
