--- Assets/Project/Scripts/NetworkingScripts/Server/PlayerDesktop.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 54ee04d6dfb06914e9b5b93f331b770d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
