--- Assets/Project/Scripts/NetworkingScripts/Server/PlayerVR.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 258fbe091bc98ea47a7ae8c3fea5766d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
