+++ Assets/Project/Scripts/NetworkingScripts/Server/ServerHandle.cs
@@ -16,13 +16,11 @@ namespace CEMSIM
                     string _username = _packet.ReadString();
 
                     Debug.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} connects successfully and whose username is {_username}");
-                    NetworkOverlayMenu.Instance.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} connects successfully and whose username is {_username}");
 
                     // check whether the packet is from the client
                     if (_clientIdCheck != _fromClient)
                     {
                         Debug.LogWarning($"Client {_fromClient} has assumed with client id {_clientIdCheck} with username {_username}");
-                        NetworkOverlayMenu.Instance.Log($"Warning: Client {_fromClient} has assumed with client id {_clientIdCheck} with username {_username}");
                         return;
                     }
 
@@ -35,13 +33,11 @@ namespace CEMSIM
                     string _msg = _packet.ReadString();
 
                     Debug.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} sends a UDP ping with msg {_msg}");
-                    NetworkOverlayMenu.Instance.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} sends a UDP ping with msg {_msg}");
 
                     // check whether the packet is from the client
                     if (_clientIdCheck != _fromClient)
                     {
                         Debug.Log($"Client {_fromClient} has assumed with client id {_clientIdCheck} ");
-                        NetworkOverlayMenu.Instance.Log($"Client {_fromClient} has assumed with client id {_clientIdCheck} ");
                         return;
                     }
 
@@ -58,7 +54,6 @@ namespace CEMSIM
                     string _msg = _packet.ReadString();
 
                     Debug.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} sends a TCP ping with msg {_msg}");
-                    NetworkOverlayMenu.Instance.Log($"Client {Server.clients[_fromClient].tcp.socket.Client.RemoteEndPoint} sends a TCP ping with msg {_msg}");
 
 
                     // Create response
@@ -75,15 +70,12 @@ namespace CEMSIM
                 /// <param name="_packet"></param>
                 public static void SpawnRequest(int _fromClient, Packet _packet)
                 {
-                    NetworkOverlayMenu.Instance.Log("RECEIVED SPAWN REQUEST");
                     string _username = _packet.ReadString();
-                    bool _vrEnabled = _packet.ReadBool();
 
                     Debug.Log($"client{_fromClient}: Spawn player.");
-                    NetworkOverlayMenu.Instance.Log($"client{_fromClient}: Spawn player.");
 
                     // send back the packet with necessary inforamation about player locations
-                    Server.clients[_fromClient].SendIntoGame(_username, _vrEnabled);
+                    Server.clients[_fromClient].SendIntoGame(_username);
                 }
 
                 /// <summary>
@@ -103,21 +95,7 @@ namespace CEMSIM
 
                     //Debug.Log($"client{_fromClient}: move packet received.");
 
-                    //Server.clients[_fromClient].player.SetInput(_inputs, _rotation);
-                    PlayerDesktop clientPlayer = (PlayerDesktop)Server.clients[_fromClient].player;
-                    clientPlayer.SetInput(_inputs, _rotation);
-                }
-
-                public static void PlayerVRMovement(int _fromClient, Packet _packet)
-                {
-                    Vector3 _position = _packet.ReadVector3();
-                    Quaternion _rotation = _packet.ReadQuaternion();
-
-                    //Debug.Log($"client{_fromClient}: move packet received.");
-
-                    //Server.clients[_fromClient].player.SetPosition(_position, _rotation);
-                    PlayerVR clientPlayer = (PlayerVR)Server.clients[_fromClient].player;
-                    clientPlayer.SetPosition(_position, _rotation);
+                    Server.clients[_fromClient].player.SetInput(_inputs, _rotation);
                 }
             }
         }
