+++ Assets/Scripts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 72856c4c27ed5d14281dfd38e057e41e
-PrefabImporter:
+guid: 7ca8b53c4960449e6b668af63add1edd
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
