+++ Assets/XR/XRGeneralSettings.asset
@@ -44,7 +44,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d2dc886499c26824283350fa532d087d, type: 3}
   m_Name: XRGeneralSettings
   m_EditorClassIdentifier: 
-  Keys: 01000000
+  Keys: 010000000d00000007000000
   Values:
   - {fileID: -7760545670062519426}
 --- !u!114 &2663332761979515837
@@ -63,6 +63,7 @@ MonoBehaviour:
   m_AutomaticLoading: 0
   m_AutomaticRunning: 0
   m_Loaders:
+  - {fileID: 11400000, guid: bce4d1085f677524c929beea7c6d38be, type: 2}
   - {fileID: 11400000, guid: a0ba533d4409886459c7f02f1a49c1aa, type: 2}
 --- !u!114 &7978545908798726749
 MonoBehaviour:
