+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,12 +11,9 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Project/Scenes/NetworkingDemo.unity
     guid: 994d12a6d69ddaa4bbfe4cb6665cc4be
-  - enabled: 0
+  - enabled: 1
     path: Assets/Project/Scenes/NetworkingDemo-Server.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
-  - enabled: 1
-    path: Assets/Project/Scenes/OperatingRoomDemo-Server.unity
-    guid: b662ef4d59b449345af9bd0840df58b7
   m_configObjects:
     Unity.XR.Oculus.Settings: {fileID: 11400000, guid: 677a0075bbe149e43af1f1cbbe1b2cea,
       type: 2}
