+++ ProjectSettings/ProjectSettings.asset
@@ -121,7 +121,10 @@ PlayerSettings:
     Others: 1
   bundleVersion: 0.1
   preloadedAssets:
-  - {fileID: 0}
+  - {fileID: 6902682411932545728, guid: bf76925e5e638374bb29b9286bfd2465, type: 2}
+  - {fileID: 7978545908798726749, guid: aa3713862873bd54787a8ba2b9377223, type: 2}
+  - {fileID: 11400000, guid: 677a0075bbe149e43af1f1cbbe1b2cea, type: 2}
+  - {fileID: 7978545908798726749, guid: aa3713862873bd54787a8ba2b9377223, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
